pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "action-knowledge-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "pydantic>=2.0",
    "alembic",
    "anthropic>=0.40.0",
    "python-dotenv",
    "typer",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["."]
